+++ .next/dev/static/chunks/app_locations_page_tsx_7ca7415e._.js
@@ -1,8 +1,8 @@
 (globalThis.TURBOPACK_CHUNK_LISTS || (globalThis.TURBOPACK_CHUNK_LISTS = [])).push({
     script: typeof document === "object" ? document.currentScript : undefined,
     chunks: [
-  "static/chunks/_012eb240._.js",
-  "static/chunks/node_modules_aef3a65d._.js"
+  "static/chunks/_89708c97._.js",
+  "static/chunks/node_modules_c1df2336._.js"
 ],
     source: "dynamic"
 });